Added post-processing to uploaded PNGs.

+++ server/modules/storage.js
@@ -10,7 +10,26 @@ var logger = require('winston');
 var File = require('mongoose').model('File');
 var config = require('./settings').current;
 
+var dimensions = {
+	backglass: [
+		{ name: 'small',   width: 253, height: 202 },
+		{ name: 'small2x', width: 506, height: 404 }
+	]
+};
+
+
 function Storage() {
+
+	// create necessary paths..
+	_.each(dimensions, function(dimTypes, dimType) {
+		_.each(dimTypes, function(dim) {
+			var p = path.resolve(config.vpdb.storage, dim.name);
+			if (!fs.existsSync(p)) {
+				logger.info('[storage] Creating non-existant %s path %s', dimType, p);
+				fs.mkdirSync(p);
+			}
+		});
+	});
 }
 
 Storage.prototype.cleanup = function(graceperiod, done) {
@@ -26,7 +45,9 @@ Storage.prototype.cleanup = function(graceperiod, done) {
 
 		async.eachSeries(files, function(file, next) {
 			logger.info('[storage] Cleanup: Removing inactive file "%s" by <%s> (%s).', file.name, file.author.email, file._id.toString());
-			fs.unlinkSync(file.getPath());
+			if (fs.existsSync(file.getPath())) {
+				fs.unlinkSync(file.getPath());
+			}
 			file.remove(next);
 		}, done);
 	});
@@ -53,8 +74,43 @@ Storage.prototype.metadata = function(file, done) {
 	}
 };
 
-Storage.prototype.url = function(obj) {
-	return '/storage/' + obj._id;
-}
+Storage.prototype.postprocess = function(file, done) {
+	var mime = file.mimeType.split('/');
+	var type = mime[0];
+	var subtype = mime[1];
+
+	var PngQuant = require('pngquant');
+	var OptiPng = require('optipng');
+
+	switch(type) {
+		case 'image':
+			if (dimensions[file.fileType]) {
+				async.eachSeries(dimensions[file.fileType], function(dimension, next) {
+
+					var quanter = new PngQuant([128]);
+					var optimizer = new OptiPng(['-o7']);
+
+					var filepath = file.getPath(dimension.name);
+					var writeStream = fs.createWriteStream(filepath);
+					writeStream.on('finish', function() {
+						logger.info('[storage] Saved quantered and optimized image to "%s".', filepath);
+						next();
+					});
+					logger.info('[storage] Optimizing "%s" for %s %s...', file.name, dimension.name, file.fileType);
+					gm(file.getPath()).resize(dimension.width, dimension.height).stream().pipe(quanter).pipe(optimizer).pipe(writeStream);
+
+				}, done);
+			} else {
+				done();
+			}
+			break;
+		default:
+			done();
+	}
+};
+
+Storage.prototype.url = function(obj, size) {
+	return obj ? '/storage/' + obj._id + (size ? '/' + size : '') : null;
+};
 
 module.exports = new Storage();
